Move files to internal
