+++ samtranslator/internal/schema_source/aws_serverless_layerversion.py
@@ -4,7 +4,7 @@ from typing import Optional, Union
 
 from typing_extensions import Literal
 
-from samtranslator.schema_source.common import (
+from samtranslator.internal.schema_source.common import (
     BaseModel,
     PassThroughProp,
     ResourceAttributes,
